// models/outmigrations.ts
export interface OutmigrationApplication {
    application_id: string;
    application_date: Date;
    country_name: string;
    verification_cadres: string;
    invoice_no: string;
    amount_due: string;
    amount_paid: string;
    balance_due: string;
    application_status: string;
    tracking_number: null;
}